+++ include/owl.h
@@ -100,9 +100,9 @@
 #define OWL_EVENT_TEXTINPUT (OWL_EVENT_BASE + 7)
 #define OWL_EVENT_TEXTEDITING (OWL_EVENT_BASE + 8)
 
-#define OWL_PI 3.14159265358979323846f
-#define OWL_DEG (180.0f / OWL_PI)
-#define OWL_RAD (OWL_PI / 180.0f)
+#define OWL_PI 3.14159265358979323846
+#define OWL_DEG (180.0 / OWL_PI)
+#define OWL_RAD (OWL_PI / 180.0)
 
 #ifdef __cplusplus
 extern "C" {
@@ -141,12 +141,12 @@ typedef struct owl_Rect {
 } owl_Rect;
 
 typedef struct owl_Vector2 {
-  f32 x, y;
+  f64 x, y;
 } owl_Vector2;
 
 typedef struct owl_Matrix {
-  f32 a, c, tx;
-  f32 b, d, ty;
+  f64 a, c, tx;
+  f64 b, d, ty;
   /*  0, 0, 1 */
 } owl_Matrix;
 
@@ -318,27 +318,27 @@ typedef enum {
 OWL_API owl_Pixel owl_rgb(u8 r, u8 g, u8 b);
 OWL_API owl_Pixel owl_rgba(u8 r, u8 g, u8 b, u8 a);
 
-OWL_API f32 owl_degrees(f32 rad);
-OWL_API f32 owl_radians(f32 deg);
+OWL_API f64 owl_degrees(f64 rad);
+OWL_API f64 owl_radians(f64 deg);
 
-OWL_API void owl_matrix(owl_Matrix *m, f32 a, f32 b, f32 c, f32 d, f32 tx,
-                        f32 ty);
+OWL_API void owl_matrix(owl_Matrix *m, f64 a, f64 b, f64 c, f64 d, f64 tx,
+                        f64 ty);
 OWL_API void owl_matrix_setidentity(owl_Matrix *m);
-OWL_API void owl_matrix_settranslate(owl_Matrix *m, f32 x, f32 y);
-OWL_API void owl_matrix_setscale(owl_Matrix *m, f32 x, f32 y);
-OWL_API void owl_matrix_setshear(owl_Matrix *m, f32 x, f32 y);
-OWL_API void owl_matrix_setrotate(owl_Matrix *m, f32 rad);
-OWL_API void owl_matrix_settransrotate(owl_Matrix *m, f32 x, f32 y, f32 rad);
+OWL_API void owl_matrix_settranslate(owl_Matrix *m, f64 x, f64 y);
+OWL_API void owl_matrix_setscale(owl_Matrix *m, f64 x, f64 y);
+OWL_API void owl_matrix_setshear(owl_Matrix *m, f64 x, f64 y);
+OWL_API void owl_matrix_setrotate(owl_Matrix *m, f64 rad);
+OWL_API void owl_matrix_settransrotate(owl_Matrix *m, f64 x, f64 y, f64 rad);
 
 OWL_API void owl_matrix_multiply(owl_Matrix *m, owl_Matrix *l, owl_Matrix *r);
 OWL_API bool owl_matrix_invert(owl_Matrix *m);
 
-OWL_API void owl_matrix_translate(owl_Matrix *m, f32 x, f32 y);
-OWL_API void owl_matrix_scale(owl_Matrix *m, f32 x, f32 y);
-OWL_API void owl_matrix_shear(owl_Matrix *m, f32 x, f32 y);
-OWL_API void owl_matrix_rotate(owl_Matrix *m, f32 rad);
-OWL_API void owl_matrix_transrotate(owl_Matrix *m, f32 x, f32 y, f32 rad);
-OWL_API void owl_matrix_apply(owl_Matrix *m, owl_Vector2 *out, f32 x, f32 y);
+OWL_API void owl_matrix_translate(owl_Matrix *m, f64 x, f64 y);
+OWL_API void owl_matrix_scale(owl_Matrix *m, f64 x, f64 y);
+OWL_API void owl_matrix_shear(owl_Matrix *m, f64 x, f64 y);
+OWL_API void owl_matrix_rotate(owl_Matrix *m, f64 rad);
+OWL_API void owl_matrix_transrotate(owl_Matrix *m, f64 x, f64 y, f64 rad);
+OWL_API void owl_matrix_apply(owl_Matrix *m, owl_Vector2 *out, f64 x, f64 y);
 
 OWL_API const char *owl_version(s32 *major, s32 *minor, s32 *patch);
